refactored so the CA is organized and I dont have my own packages
