+++ src/main/java/data_access/FileUserDataAccessObject.java
@@ -15,8 +15,7 @@ import java.util.Objects;
 import entities.Property;
 import entities.User;
 import entities.UserFactory;
-import use_case_rosa.universities.UniversitiesDataAccessInterface;
-import use_case_rosa.universities.UniversitiesUserDataAccessInterface;
+import usecases.universities.UniversitiesUserDataAccessInterface;
 import usecases.change_password.ChangePasswordUserDataAccessInterface;
 import usecases.itinerary.ItineraryDataAccessInterface;
 import usecases.login.LoginUserDataAccessInterface;
