+++ src/main/java/interface_adapters/universities/UniversitiesViewModel.java
@@ -1,11 +1,8 @@
-package interface_adapter_rosa.universities;
+package interface_adapters.universities;
 
-import entity_rosa.University;
-
-import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
 
-public class UniversitiesViewModel extends ViewModel<UniversitiesState>{
+public class UniversitiesViewModel extends ViewModel<UniversitiesState> {
     private final PropertyChangeSupport support = new PropertyChangeSupport(this);
 
     public UniversitiesViewModel(){
