+++ src/main/java/usecases/universities/UniversitiesUserDataAccessInterface.java
@@ -1,7 +1,6 @@
-package use_case_rosa.universities;
+package usecases.universities;
 
 import entities.User;
-import entity_rosa.University;
 
 public interface UniversitiesUserDataAccessInterface {
     User get(String username); // returns the current user based on the username
