+++ src/main/java/view/DashboardView.java
@@ -2,11 +2,9 @@ package view;
 
 import app.FlightUseCaseFactory;
 import app.PropertyUseCaseFactory;
-import entities.Property;
-import entity_rosa.University;
-import interface_adapter_rosa.universities.*;
-import interface_adapter_rosa.universities.UniversitiesController;
-import interface_adapter_rosa.universities.UniversitiesViewModel;
+import interface_adapters.universities.UniversitiesController;
+import interface_adapters.universities.UniversitiesState;
+import interface_adapters.universities.UniversitiesViewModel;
 import java.awt.Component;
 
 import javax.swing.BoxLayout;
@@ -20,20 +18,15 @@ import interface_adapters.flight.FlightViewModel;
 import interface_adapters.itinerary.ItineraryController;
 import interface_adapters.itinerary.ItineraryViewModel;
 import interface_adapters.logged_in.LoggedInState;
-import interface_adapters.property.PropertySelectedCallback;
 import interface_adapters.property.PropertyState;
 import interface_adapters.property.PropertyViewModel;
 import usecases.flight.FlightUserDataAccessInterface;
 import usecases.itinerary.ItineraryDataAccessInterface;
 import usecases.property.PropertyUserDataAccessInterface;
-import use_case_rosa.universities.UniversitiesUserDataAccessInterface; // rosa import added
-import view_rosa.UniversitiesView;
-import entity_rosa.University;
+import usecases.universities.UniversitiesUserDataAccessInterface; // rosa import added
+
 import javax.swing.*;
 import java.awt.*;
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
-import java.util.List;
 
 import static javax.swing.JOptionPane.QUESTION_MESSAGE;
